+++ Sources/BraintreeCard/Public/BraintreeCard/BraintreeCard.h
@@ -8,10 +8,6 @@ FOUNDATION_EXPORT const unsigned char BraintreeCardVersionString[];
 
 #if __has_include(<Braintree/BraintreeCard.h>)
 #import <Braintree/BTCardClient.h>
-#import <Braintree/BTCard.h>
-#import <Braintree/BTCardNonce.h>
 #else
 #import <BraintreeCard/BTCardClient.h>
-#import <BraintreeCard/BTCard.h>
-#import <BraintreeCard/BTCardNonce.h>
 #endif